WIP: Fix AdminTopPostDto Long type, update api_config to 10.0.2.2 for Android emulator

+++ spring_boot_backend/src/main/kotlin/com/androidinsta/DTO/AdminDto.kt
@@ -139,7 +139,7 @@ data class AdminTopPostDto(
     val caption: String?,
     val username: String,
     val userId: Long,
-    val likesCount: Int,
-    val commentsCount: Int,
+    val likesCount: Long,
+    val commentsCount: Long,
     val createdAt: LocalDateTime
 )
